Load language-specific modules earlier
- Language-specific modules requested in the \usepackage{xcref} line are
  now loaded when \ProcessKeysOptions { xcref } is executed, which is
  (now) at the very end of xcref.sty.

- For other cases (i.e., when using \xcrefusemodules & friends), they
  are immediately loaded.

This removes the need for \AtBeginDocument wrapping in the document
preamble when doing things such as \let\fxcref=\xcreffrenchwrapper. This
change is also needed for the next commit, because it will add code that
needs to access functions defined in a language-specific module and must
be executed in the preamble.