Introduce \xcref@tabacckludge
This is a \protected variant of \@tabacckludge. We need it because for
instance, \tl_mixed_case:nn { ... } { \@tabacckludge`a } doesn't work as
expected[1], due to the fact that \tl_mixed_case:nn performs recursive
expansion on its second argument before changing the case. Replace all
uses of \@tabacckludge with \xcref@tabacckludge in xcref.sty and its
language modules.

[1] Tested with ... = french, which should have no particular effect as
    it does not belong to the set of special values recognized by
    \tl_mixed_case:nn (yes, I know the special values are language codes
    such as 'nl' or 'de-alt', but this doesn't matter here: 'french' is
    what with xcref passes to \tl_mixed_case:nn as its first argument
    when the selected language is 'french', because there is no special
    code for French such as 'fr' that is currently documented in
    interface3.pdf, thus we simply keep the babel language name in such
    cases, at least for now).