\xcref:nn: restrict setup to types used in #2; call @crefdefineallformats
- In \xcref:nn, only call \crefname and \Crefname for the needed types
  (those in the second argument). \g_@@_french_types_seq and
  \g_@@_ngerman_types_seq are not really used anymore (they just provide
  records of the registered types).

- At the very end of the language-specific "mapfunc"
  (currently, \@@_french_setup_cref_names_mapfunc:n and
  \@@_ngerman_setup_cref_names_mapfunc:n), after the \crefname and
  \Crefname calls, use \@crefdefineallformats to make sure cleveref
  updates all macros that may be needed later, such as
  \cref@<type>@format@first.

The first item makes things faster, the second slower. The real gain is
that thanks to the use of \@crefdefineallformats, it is not necessary
anymore to keep \crefname and \Crefname calls in the preamble that are
redundant with xcref data (e.g., data provided with
\xcrefngermanSetNamesTableEntry or any of the language-specific
'/xcref/<language>/names table' options)---that is, as long as you don't
use cleveref directly.