Remove irrelevant entries from the index
Many entries were just LaTeX3 functions used in the implementation or in
its documentation. Most of these added no real value to the index, but
were making it unnecessarily hard to find relevant material in the
index. As for \l_@@_ngerman_article_and_prep_prop, the problem was
different: it was used in the documentation before the @@ prefix had
been defined, which resulted in an index entry for
\l_@@_ngerman_article_and_prep_prop (literally).