Test files: remove formatting in a command used for PDF bookmarks
There was a little bug in \__xcref_test_dumbsectioningcmd:nn. It is not
used in xcref-based documents (it's only used in test files) and wasn't
used anywhere in a way that triggers the bug, though.

There is probably a simpler way using \pdfstringdefDisableCommands, but
I want to get this right for history anyway.