Add code for documenting PGF keys within l3doc
Add pgfkey and pgfkey* environments supporting a syntax close to
tcolorbox's docKey and docKey* environments (except for options). The
implementation is probably different---I haven't looked how docKey and
docKey* are implemented in tcolorbox. These environments automatically
create a label that can be referenced with command \pgfkeyRef, also
added in this commit.

The pgfkey environment adds the key to the index, in two places: one
under letter P for "PGF keys", the other under the first character of
the key path following the leading slash. On the other hand, The pgfkey*
environment doesn't write any index entry (this is the same distinction
as for docKey and docKey*).

In the index, keys are formatted using \verb* as per l3doc's
\__codedoc_special_index_set:Nn. On the other hand, the pgfkey and
pgfkey* environments, as well as the \pgfkeyRef command introduced here,
format keys using \__xcrefdoc_pgfkey_verb_like:n, which uses \scantokens
to retokenize its argument under a category code régime which is the
same as the one set by the LaTeX kernel verbatim environments (we use
\dospecials). The point of using \scantokens here---which verbatim from
the LaTeX kernel doesn't do---is that it allows one to easily build the
argument of \__xcrefdoc_pgfkey_verb_like:n (possibly using variants),
write it as part of the replacement text of a macro, etc.

When typesetting key paths, we also disable ligatures with \@noligs as
done for verbatim text, in case some key paths were to contain things
such as ",,", "<<" or "--". Not very likely, but it's better this way.